Introduce stat to track when we skip samples at the end of "Waiting for GVL"
This is useful because our main control for "Waiting for GVL" is the
waiting_for_gvl_threshold_ns, and thus we can use this stat to monitor
how many events are being skipped as being under that threshold.

Note that regular (periodic) samples can still observe the
"Waiting for GVL" state; but there will not be any special samples
created to represent the beginning or end of "Waiting for GVL".