Remove `pending_update` mechanism from Heap recorder
To simplify the heap cleanup after GC PR, I'm changing the mechanism
to become best-effort, e.g. something that exists to reduce memory and
(hopefully) serialization latency, but that is not required for
correctness.

As the mechanism becomes best-effort, we no longer need to track
that we missed an update.