Add lib injection support for Ruby 3.3
Support for Ruby 3.3 is present in the gem, the only thing needed for injection is to change the installation script to allow it to run with that version of Ruby.

Because the Ruby API does not change in incompatible ways in between patch versions, we declare our compatibility with the API version "3.3.0", thus supporting all 3.3.x versions.

Signed-off-by: Marco Costa <marco.costa@datadoghq.com>