Fix segfaults when calling `_native_debug_heap_recorder` in tests
When the heap recorder was tracking internal VM objects, because such
objects have no klass, they cannot be inspected using regular APIs.

Trying to do that results in a segfault as Ruby does not check before
trying to access that info.

To work around the issue, make sure to use the VM special object
printing API for debugging (`rb_obj_info`) which handles these
cases fine.

(Note that this issue was only for tests, because we only attempt
to print object information in this situation.)