[PROF-10589] Add benchmarking configuration for gvl profiling
**What does this PR do?**

This PR tweaks our nightly benchmark runs configuration to add a new
configuration that enables GVL profiling.

**Motivation:**

This will allow us to have better coverage for testing this feature.

**Additional Notes:**

This was previously blocked on upgrading the gitlab image that
we run in the benchmarking platform, which I finally took care
of in https://github.com/DataDog/benchmarking-platform/pull/108 .

**How to test the change?**

I'll trigger a manual run to validate it's running fine and report
back ;)