[NO-TICKET] Fix crashtracking sending parent pid/runtime-id when child crashes
**What does this PR do?**

This PR fixes a bug in the crashtracking component: because it cached
all of the tags from creation, if a process forked, it would use the
tags from the parent, which included a (now-incorrect) pid and
runtime-id.

**Motivation:**

This makes sure we have accurate metadata in crashtracking reports.

**Additional Notes:**

N/A

**How to test the change?**

I've added test coverage for this change. It can also be observed by
using

```
$ DD_CRASHTRACKING_ENABLED=true bundle exec ruby -e 'require "datadog/auto_instrument"; puts "Parent: #{Process.pid} #{Datadog::Core::Environment::Identity.id}"; fork { puts "Child: #{Process.pid} #{Datadog::Core::Environment::Identity.id}"; Process.kill("SEGV", Process.pid) }'
```

and validating that the correct pid/runtime-id from the child
is sent in the metadata.