[NO-TICKET] Fix rubocop lint
**What does this PR do?**

This PR moves the `start_worker` and `stop_worker` to the `private`
section of `writer.rb`.

It does this instead of using
`private :start_worker, :stop_worker` to make rubocop happy.

**Motivation:**

Unbreak rubocop in master.

**Additional Notes:**

I thought we had pinned down the version of rubocop in CI, but it
appears we have a gap. I've not investigated further.

**How to test the change?**

Validate that the lint check in CI is back to green.