[NO-TICKET] Upgrade to libdatadog 14.0
**What does this PR do?**

This PR upgrades the libdatadog dependency to the latest 14.0 release.

It also updates our API usage to match some 13.1 -> 14.0 changes.

**Motivation:**

The libdatadog 14.0 release includes a number of improvements to
the crashtracking implementation, and we want to pick them up.

**Additional Notes:**

As a first pass, I've disabled the creation and use of the alt
stack, to match our existing setup.

I did not yet experiment with `use_alt_stack = true`.

**How to test the change?**

Existing test coverage should be enough to cover our libdatadog
API usage.