Enable `use_alt_stack` setting for crashtracker
I spent some time looking into `use_alt_stack` setting and I think
there's a bug in libdatadog and that setting doesn't seem to be doing
anything at this point.

Since the Ruby VM itself sets an altstack, this leaves us:

* In libdatadog 13.1, the crashtracker code always used the altstack
  (e.g. equivalent of `use_alt_stack` was hardcoded to `true`)
  => in libdatadog 13.1 we used the Ruby altstack
* In libdatadog 14.0, the crashtracker code chooses to use the altstack
  based on the value of `create_alt_stack` => libdatadog 14.0 we can't
  use the Ruby altstack

In preparation for `use_alt_stack` to get fixed in a future libdatadog
release, I'm changing it to true although, right now it's a no-op.

This means that in a future libdatadog release, we'll be back to where
we were in libdatadog 13.1: the crashtracker DOES NOT create its
own altstack, BUT WILL USE the existing Ruby altstack.