[NO-TICKET] Relax profiling CODEOWNERS configuration
**What does this PR do?**

This PR relaxes the GitHub `CODEOWNERS` file rules for things related
to profiling.

Specifically, it adds `@DataDog/ruby-guild` as a second owner on
top of just `@DataDog/profiling-rb`.

**Motivation:**

Now that we've enableed the "Require review from Code Owners" feature
on GitHub, since there's not a lot of us working on profiling, it
hasn't worked very well.

Specifically, for very small PRs (#4072 is a good example where we
slightly tweaked a test to avoid flakiness), it's annoying to have
to wait for a very small reviewer pool, when most folks that
are on the Ruby guild can help out review such changes.

**Additional Notes:**

N/A

**How to test the change?**

Check that future PRs affecting profiling can be merged with a
review from anyone on the ruby-guild team.