[NO-TICKET] Re-enable ASAN memory leaks testing
**What does this PR do?**

This PR re-enables the memory leak testing using the asan tool job in
CI.

For context, we added this testing in #3864 (more details about what
asan is can be found in that PR), but had to disable it in #3915
as the upstream Ruby builds we were using were broken and unavailable
for a while.

This has since been fixed upstream, and so let's re-enable this CI job.

In fact, upstream now officially allows these builds to be used, so
we no longer even need our fork of `setup-ruby` to enable them.

**Motivation:**

Improve our testing for native memory leaks.

**Additional Notes:**

N/A

**How to test the change?**

Validate that the "test-asan" is now running in CI.