DEBUG-2334 custom ActiveRecord model serialization (#4088)
Adds a custom serializer for ActiveRecord models
Adds serializers for Time, Date and DateTime (needed for ActiveRecord timestamps)

Motivation:

AR models have many fields that are not interesting for DI, and serializing them would 1) consume a lot of resources 2) cause the serialized payload to exceed the limits of the backend 3) not serialize actual attributes due to depth limits (actual attributes are already 3 levels down from the AR model). The custom serializer currently only serializes the attributes, though additional fields can be added in the future such as the new_record? flag.