[PROF-10967] Fix profiler not loading due to "rb_obj_info" symbol not found
**What does this PR do?**

This PR fixes a bug introduced in #4020, specifically in f581076c0f927 .

We started using the `rb_obj_info` to print debug information about
objects in some cases, BUT I failed to notice that this API is not
really available on Ruby 2.5 and 3.3 (but is on all others, which
is why it tripped me).

This manifested in the following error reported by a customer:

> WARN -- datadog: [datadog] Profiling was requested but is not
> supported, profiling disabled: There was an error loading the
> profiling native extension due to 'RuntimeError Failure to load
> datadog_profiling_native_extension.3.3.5_x86_64-linux-musl due
> to Error relocating
> /app/vendor/bundle/ruby/3.3.0/gems/datadog-2.6.0/lib/datadog/profiling/../../datadog_profiling_native_extension.3.3.5_x86_64-linux-musl.so:
> rb_obj_info: symbol not found' at
> '/app/vendor/bundle/ruby/3.3.0/gems/datadog-2.6.0/lib/datadog/profiling/load_native_extension.rb:41:in `<main>''

This PR fixes this issue by never referencing `rb_obj_info` on
those Rubies. Since this API is only used for printing information
during errors, this should be fine (and is better than the alternative
of not printing info on any Rubies).

**Motivation:**

Fix profiling not loading in certain situations on Ruby 2.5 and 3.3.

**Additional Notes:**

Interestingly, this issue did not show up on glibc systems. I guess
musl libc is being a bit more eager about trying to resolve symbols?

**How to test the change?**

This change includes test coverage. Disabling the added check
in `extconf.rb` will produce a failing test.