Remove leftovers of class name gathering being optional
At some point in the past, class name gathering was optional. This is
now removed, but we still had this vestigial `optional_class_name`
variable + passing a pointer to it, rather than passing the `class_name`
as a value.

This commit cleans this up and makes all APIs require/assume there's a
`class_name` for an allocated object.

(This enables a few other cleanups I want to make next.)