[PROF-10987] Add Ruby GC guard to heap profiler as a preventative measure
**What does this PR do?**

This PR adds a `RUBY_GC_GUARD` call to the Ruby heap profiler, making
sure that the object being pointed by the `ref` local variable does
not get garbage collected too early.

**Motivation:**

This is a speculative fix:
We've seen a crash coming from calls to `ruby_obj_memsize_of`
(line 654). Adding this `RUBY_GC_GUARD` is a very-low-cost
"just in case" until we can fully get to the bottom of this crash,
and even if it wasn't needed, it's very harmless to have.

**Additional Notes:**

N/A

**How to test the change?**

This is part of the challenge -- it's extremely hard to test this
change, as in many compilers it may even turn out to be a no-op.

We'll need to wait for more info on the crash to go further than
this change.