Introduce `is_safe_to_allocate_objects` to avoid unsafe allocations
When called from inside an allocation (NEWOBJ tracepoint), it's not
safe to allocate further new objects (including exceptions to be
raised).

One such example of an allocation is when calling
`read_otel_current_span_key_const` to initialize the otel span key.

Thus, this commit introduces the `is_safe_to_allocate_objects`
flag (and plumbs it around a bunch of methods...) so that we can
gate calls to `read_otel_current_span_key_const` and not perform
them when they're not safe.