[PROF-11007] Upgrade to libdatadog 14.3.1
**What does this PR do?**

This PR upgrades the datadog gem to use libdatadog 14.3.1.

There were no needed API changes, it's a drop-in upgrade.

**Motivation:**

Staying on the latest libdatadog enables us to pick up the latest
improvements and also makes it easier to develop new features that need
libdatadog changes.

**Additional Notes:**

As usual, I'm opening this PR as a draft as libdatadog 14.3.1 is not yet
available on rubygems.org, and I'll come back to re-trigger CI and mark
this as non-draft once it is.

**How to test the change?**

Our existing test coverage includes libdatadog testing, so a green CI is good here :)