[NO-TICKET] Re-add at_fork hook testing for crashtracking
**What does this PR do?**

This PR re-adds the spec that validates that the at_fork hook in
crashtracking correctly picks up the latest instance of the
component.

**Motivation:**

Since this is quite a weird corner case, I think it's important to
have a spec to cover it.

**Additional Notes:**

This spec was temporarily removed in #3970 with the intention of
re-adding it in #3983 but we kinda forgot about it.

**How to test the change?**

Check this test passes as-is, and fails if you change the
`update_on_fork` call to use `self`, for instance.