Introduce SignalHandlerSamplingBuffer
The SignalHandlerSamplingBuffer is, on purpose, very similar to the
sampling_buffer.

The key difference is that it will get written to in the profiler signal
handler (as the name indicates) by whatever thread is active, and thus
it needs to be an actual Ruby object so it can do marking + it needs
a few different fields.