Wrap info logs in profiling component with an OnlyOnce
This makes sure these information messages, when printed, are not
printed more than once.

I considered what to do with the warning messages as well, but since
having individual OnlyOnce instances is quite a bit of boilerplate
for the number of warnings we are emitting, I decided to target only
this one for now.