[NO-TICKET] Update crashtracker with libdatadog 15 breaking changes
**What does this PR do?**

This PR updates the crasktracker C code to build with the latest
libdatadog changes (in main) that will become part of libdatadog 15.

**Motivation:**

I'm working on a libdatadog branch and had to do this to unblock
my work, so I decided I'll create a PR with it so nobody needs to
repeat this work.

**Additional Notes:**

I'm opening this PR as draft as we shouldn't merge this until
libdatadog 15 is out.

**How to test the change?**

Existing test coverage is enough to validate this.