[NO-TICKET] Disable flaky barrier spec
**What does this PR do?**

This PR disables a spec that's been flaky for a very long time
(see https://github.com/DataDog/ruby-guild/issues/191, but I suspect
I saw it back in https://github.com/DataDog/ruby-guild/issues/50#issuecomment-1874210356 )
and it just failed on me again:
https://github.com/DataDog/dd-trace-rb/actions/runs/12706052002/job/35418238631?pr=4234

I'm not sure this test is getting us a lot of value, so let's stop
the flaky impact until we can dedicate time to fixing it.

(Maybe @lloeki can help here as the original author of the class and
tests?)

**Motivation:**

Get to zero flaky tests in our codebase!

**Additional Notes:**

N/A

**How to test the change?**

Validate spec is now skipped.