Simplify comments around StackRecorder initialization
A lot of these comments were leftovers from when the last thing we did
in `_native_new` was the call to `TypedData_Wrap_Struct`.

Refactoring work done over time, as well as evolution of the libdatadog
APIs meant we could allocate the Ruby object sooner, and thus a lot
of these comments became not relevant.