Intern strings in string storage only once
We don't actually need to keep re-interning these strings in the
current version of the string storage; I suspect this was a leftover
of the time when the string storage automatically cleaned up strings
that were not used in a certain generation, but this is no longer
the case.

In fact, now that libdatadog relies on counting intern/unintern
calls to know which strings to keep, calling `intern` forever on
these strings would eventually lead to an overflow, and libdatadog
would start returning errors on the calls to `intern`.

(In turn, on our side, we'd raise the exception and thus stop
profiling which is not amazing)