Improve stack comparison and hashing, taking advantage of it being only ids now
With the move to interned ids in libdatadog, the same stacks are
represented by the same ids, which means the same underlying bytes.

Thus, rather than having per-item comparisons and whatnot (which we
needed because before we had pointers to strings, and strings were not
interned/unique), we can now compare and hash the whole stack at once
as a binary blob.

This is crucially missing tests! I'll come back to add them, we
definitely want some test coverage on this key element.