Minor: Relax profiling assertion to avoid flakiness
In CI we ran into a flaky failure from this spec:
<https://app.circleci.com/pipelines/github/DataDog/dd-trace-rb/18799/workflows/001aa960-917a-4036-83e1-9965bdf5ff34/jobs/659509>

```
Failures:

  1) Datadog::Profiling::Collectors::ThreadContext#sample timeline support when timeline is enabled when thread starts Waiting for GVL cpu-time behavior on Linux assigns all the cpu-time to the sample before Waiting for GVL started
     Failure/Error: expect(first_sample.values.fetch(:"cpu-time")).to be 12345

       expected #<Integer:24691> => 12345
            got #<Integer:44213> => 22106

       Compared using equal?, which compares object identity,
       but expected and actual are not the same object. Use
       `expect(actual).to eq(expected)` if you don't care about
       object identity in this example.
     # ./spec/datadog/profiling/collectors/thread_context_spec.rb:1201:in `block (7 levels) in <top (required)>'
     # ./spec/spec_helper.rb:238:in `block (2 levels) in <top (required)>'
     # ./spec/spec_helper.rb:123:in `block (2 levels) in <top (required)>'
     # /usr/local/bundle/gems/webmock-3.23.0/lib/webmock/rspec.rb:39:in `block (2 levels) in <top (required)>'
     # /usr/local/bundle/gems/rspec-wait-0.0.9/lib/rspec/wait.rb:46:in `block (2 levels) in <top (required)>'

Finished in 34.85 seconds (files took 1.67 seconds to load)
767 examples, 1 failure, 20 pending

Failed examples:

rspec ./spec/datadog/profiling/collectors/thread_context_spec.rb:1196 # Datadog::Profiling::Collectors::ThreadContext#sample timeline support when timeline is enabled when thread starts Waiting for GVL cpu-time behavior on Linux assigns all the cpu-time to the sample before Waiting for GVL started
```

This means that we observed more cpu-time than expected for the
target thread. In theory, the thread should be completely stopped
(since it's sleeping), so it should not have any cpu-time.

In practice, I can imagine that it's not impossible to make the
thread wake up spuriously -- to handle a signal, or it was woken
up on some internal Ruby lock -- just for it to go back to sleep.

The intention of this spec is to validate that all cpu-time
is assigned to the first sample (before the waiting for GVL
started), and none to the second, so I think we're OK with
not coupling to an exact amount of cpu-time to expect.