[PROF-11306] Upgrade libdatadog dependency to 16.0.1
**What does this PR do?**

This PR upgrades the datadog gem to use libdatadog 16.0.1.

It includes a few changes to match breaking API updates in
crashtracking.

**Motivation:**

Libdatadog 16 is needed to unblock
https://github.com/DataDog/dd-trace-rb/pull/4331 .

This version also brings a few crashtracking improvements.

**Change log entry**

Yes. Upgrade libdatadog dependency to 16.0.1

**Additional Notes:**

As usual, I'm opening this PR as a draft as libdatadog 16.0.1 is not
yet available on rubygems.org, and I'll come back to re-trigger CI
and mark this as non-draft once it is.

**How to test the change?**

Our existing test coverage includes libdatadog testing, so a green C
is good here :)