[PROF-9476] Revert "Temporary disable benchmark to make CI happy"
**What does this PR do?**

This reverts commit 569bf568b80e9615027a8d0c41a8a941deeb9981.

**Motivation:**

We temporarily disabled this benchmark in
https://github.com/DataDog/dd-trace-rb/pull/4331
for the reasons explained in the comment.

Once that PR is in master we can re-enable the benchmark.

**Additional Notes:**

This sharp edge of our benchmarks setup is annoying, but I'll fight
it another day...

**How to test the change?**

Validate this benchmark is now running in the benchmarking
platform results page.