Increase default heap sampling rate to 1
Given our current benchmark results, we've decided to use some of the
headroom we gained with the managed string storage branch to improve
heap sampling results.

Thus, we'll take a heap sample for every object that the allocation
profiler samples, rather than every 10th object.

We still expose this control -- via both code and environment variable,
so it can be reset back to 10 on a case-by-case basis.

BUT if you're considering raising this value, please let us know about
it -- we'd love to know your use-case.