[PROF-11151] Fix test-memory-leaks flaky behavior
**What does this PR do?**

This PR will hopefully fix almost [all of the the
flakiness](https://github.com/DataDog/dd-trace-rb/pulls?q=is%3Apr+asan)
we've seen from the test-memory-leaks GitHub workflow.

The previous "asan" builds we were using were built from ruby-head,
which means that any instability or early breakage in ruby-head would
make test-memory-leaks fail.

To fix this, I've worked with upstream to create these 3.4-asan builds:
these are Ruby builds that are built **from the latest 3.4 stable Ruby**
with asan. Thus any breakages we see in them, should also exist in
regular 3.4 builds.

**Motivation:**

With this change, the test-memory-leaks workflow becomes a lot more
valuable, since it's now never expected to fail.

Thus, any failures we see in it are worth investigating.

**Additional Notes:**

For context, asan (or ASan) is the "AddressSanitizer" tool, see
https://github.com/google/sanitizers/wiki/AddressSanitizer for more
details.

**How to test the change?**

Validate that the updated workflow is running on ruby-3.4.2 and that
it still has the diagnostic output from asan.