[PROF-11412] Add new otel appraisal variant for profiling
**What does this PR do?**

This PR adds a new otel appraisal variant for profiling that uses
`opentelemetry-api` >= 1.5 .

This is because 1.5 includes
https://github.com/open-telemetry/opentelemetry-ruby/pull/1807
and we'll need to add special support for it.

**Motivation:**

I'm adding already the appraisal version (disabled for now) to
avoid any future conflicts with changes to the `Matrixfile`.

**Additional Notes:**

N/A

**How to test the change?**

This new group is not yet used -- green CI is enough for this PR.