[PROF-11412] Support correlating profiling with otel-api 1.5+
**What does this PR do?**

This PR adds support for correlating profiling wih otel-api 1.5+.

Context storage was moved in
https://github.com/open-telemetry/opentelemetry-ruby/pull/1807
and we needed to update the profiler to account for this.

**Motivation:**

Keep profiling + otel correlation working.

**Additional Notes:**

N/A

**How to test the change?**

In #4421 I had already bootstrapped the new appraisal groups needed
for testing this. This PR enables them, and our existing test
coverage will cover the new code path when used with otel-api 1.5+.