[PROF-11405] Revert "Temporary disable benchmark to make CI happy", second try
**What does this PR do?**

This reverts commit 569bf568b80e9615027a8d0c41a8a941deeb9981.

We had previously reverted that commit in
https://github.com/DataDog/dd-trace-rb/pull/4376 but due to a bug
had to undo all the heap profiling-related changes in
https://github.com/DataDog/dd-trace-rb/pull/4409 .

After https://github.com/DataDog/dd-trace-rb/pull/4460 gets merged,
we can finally enable this benchmark again.

**Motivation:**

The benchmark was disabled for the reason mentioned in the comment.

**Additional Notes:**

I'm going to open this PR stacked on top of
https://github.com/DataDog/dd-trace-rb/pull/4460, but this PR should
only be merged after that other PR is merged to master.

**How to test the change?**

Validate this benchmark is now running in the benchmarking platform
results page.