[PROF-11475] Add extra safety checks when getting line number in profiler
**What does this PR do?**

This PR adds a few extra checks when getting the line number during
stack trace gathering in the profiler.

We got one automated report of a crash in the `rb_iseq_line_no`
function, and we couldn't reproduce it, so this is the first attempt
at avoiding the issue -- let's see if we ever observe this crash again
or not.

**Motivation:**

Improve stability of the profiler.

**Additional Notes:**

See the internal PROF-11475 ticket for more details.

**How to test the change?**

Our existing test coverage already validates that stack trace gathering
works fine and matches the Ruby APIs. If this condition was not correct,
we'd get more lines set to 0, and our unit tests would fail. (I did some
quick manual testing to check this was the case -- and indeed it is).