Introduce `EncodedProfile` class to wrap profiling data
Instead of returning an encoded profile as a bunch of bytes, let's use
a new `EncodedProfile` class to contain this data. For now, we
immediately turn the `EncodedProfile` into a string again once
we get to `stack_recorder.rb`.

This is the first step in a bigger refactoring:
The objective of this refactoring is to get the `EncodedProfile`
instance itself to the `Profiling::HttpTransport`, as starting with
libdatadog 17.0, it takes in directly an `ddog_prof_EncodedProfile`
instead of the profile bytes.